# lib/api-zod/src/generated/types/codingProblem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CodingProblem {
  id: number;
  slug: string;
  title: string;
  statement: string;
  difficulty: string;
  topic: string;
  supportedLanguages: string[];
  examples: string;
  constraints: string;
  expectedApproach: string;
}