# lib/api-zod/src/generated/types/codingSubmissionInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CodingSubmissionInput {
  problemId: number;
  language: string;
  /** @maxLength 20000 */
  sourceCode: string;
}