# lib/api-zod/src/generated/types/codingSubmissionResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CodingSubmissionResult {
  accepted: boolean;
  status: string;
  message: string;
}