# lib/api-zod/src/generated/types/engineeringMaterial.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EngineeringMaterial {
  id: number;
  branch: string;
  subject: string;
  title: string;
  materialType: string;
  summary: string;
  topics: string[];
}