# lib/api-zod/src/generated/types/feedback.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Feedback {
  id: number;
  category: string;
  subject: string;
  description: string;
  status: string;
  createdAt: string;
}