# lib/api-zod/src/generated/types/feedbackInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FeedbackInput {
  /** @maxLength 80 */
  category: string;
  /** @maxLength 160 */
  subject: string;
  /** @maxLength 5000 */
  description: string;
  /**
     * @maxLength 500
     * @nullable
     */
  attachmentUrl?: string | null;
}