# lib/api-zod/src/generated/types/interviewQuestion.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InterviewQuestion {
  id: number;
  role: string;
  category: string;
  prompt: string;
  difficulty: string;
  answerGuide: string;
}