# lib/api-zod/src/generated/types/job.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Job {
  id: number;
  title: string;
  company: string;
  location: string;
  workType: string;
  experienceRequired: string;
  requiredSkills: string[];
  /** @nullable */
  salary: string | null;
  description: string;
  qualifications: string;
  /** @nullable */
  applicationDeadline: string | null;
  sourceUrl: string;
  updatedAt: string;
}